Refactored the code to have separate class
